Changed Search generation mechanism and added a new SearchGenerator interface (#60)
Closes #58 

A new search generator was added, `FakerSearchGenerator`. As the name
implies, it uses the library `Faker` to both generate a random time to
sleep, and a valid sentence to search of 6 words (more or less; see
`variable_nb_words=True`).

Also, now the characters are not sent to the input text field at once,
but they are sent one by one with `selenium.ActionChains`. This is still
compatible with the old `RandomSearchGenerator`, since also a backspace
is a valid key to send to a text field.

Another change is the time to sleep; it is now a generator too, so its
value can be changed based on some requirements.